add curency converter

+++ Models/Balance.cs
@@ -3,12 +3,12 @@
     public class Balance
     {
         public int id { get; set; }
-        public long balance { get; set; }
+        public double balance { get; set; }
     }
     public class TransferBalance
     {
-        public int from { get; set; } 
-        public int to { get; set; } 
+        public int from { get; set; }
+        public int to { get; set; }
         public long moneyAmount { get; set; }
     }
 }
